# Property injection in constructor