--- a/src/index.ts
+++ b/src/index.ts
@@ -1,9 +1,9 @@
-import { InvocationHandlerWrapper } from "./invocation-handler-wrapper";
+import { InvocationHandlerWrapper } from './invocation-handler-wrapper';
 
 export function configure(frameworkConfiguration) {
-    frameworkConfiguration.container.setHandlerCreatedCallback((handler) => {
-        return new InvocationHandlerWrapper(handler.fn, handler.invoker, handler.dependencies);
-    });
+    frameworkConfiguration.container.setHandlerCreatedCallback(handler =>
+        new InvocationHandlerWrapper(handler.fn, handler.invoker, handler.dependencies)
+    );
 }
 
 export {InvocationHandlerWrapper};