--- a/src/index.ts
+++ b/src/index.ts
@@ -5,7 +5,7 @@ export * from './decorators';
 export function configure(frameworkConfiguration) {
     frameworkConfiguration.container.setHandlerCreatedCallback(handler =>
         (<any>handler.fn).injectProperties ?
-          new InvocationHandler(handler.fn, handler.invoker, handler.dependencies) :
+          new InvocationHandler(handler.fn, null, handler.dependencies) :
           handler
     );
 }