--- a/src/invocation-handler-wrapper.ts
+++ b/src/invocation-handler-wrapper.ts
@@ -1,4 +1,4 @@
-import { InvocationHandler } from "aurelia-dependency-injection";
+import { Container, InvocationHandler } from 'aurelia-dependency-injection';
 
 /**
 * Wraps the core `InvocationHandler` to inject into properties.
@@ -11,28 +11,35 @@ export class InvocationHandlerWrapper extends InvocationHandler {
     * @param dynamicDependencies Additional dependencies to use during invocation.
     * @return The result of the function invocation.
     */
-    invoke(container, dynamicDependencies) {
-        let instance = super.invoke(container, dynamicDependencies);
-        return this.injectProperties(container, instance);
+    invoke(container: Container, dynamicDependencies?: any[]): any {
+        return (<any>this.fn).injectProperties ?
+                invoke(container, this.fn, this.dependencies, dynamicDependencies) :
+                super.invoke(container, dynamicDependencies)
     }
+}
 
-    /**
-    * Injects property dependencies if the conventional `injectProperties` is defined.
-    * @param container The calling container.
-    * @param instance The target of injection.
-    * @return The instance with injected properties.
-    */
-    injectProperties(container, instance) {
-        if ("injectProperties" in this.fn) {
-            let dependencies = this.fn["injectProperties"];
-            for (let property in dependencies) {
-                instance[property] = container.get(dependencies[property]);
-            }
-            if ("afterConstructor" in instance) {
-                instance.afterConstructor.call(instance);
-            }
-        }
-        return instance;
-    }
+function invoke (container, fn, staticDependencies, dynamicDependencies) {
+  const injectProperties = fn.injectProperties;
+  const injectProps = Object.create(null);
+  for (let property in injectProperties) {
+    injectProps[property] = {
+      value: container.get(injectProperties[property])
+    };
+  }
+  // inject properties as soon as object is created, before calling constructor
+  let instance = Object.create(fn.prototype || null, injectProps);
+  let i = staticDependencies.length;
+  let args = new Array(i);
 
+  while (i--) {
+    args[i] = container.get(staticDependencies[i]);
+  }
+
+  if (dynamicDependencies) {
+    args = args.concat(dynamicDependencies);
+  }
+
+  // call constructor
+  fn.apply(instance, args);
+  return instance;
 }