This is something I am ready to release as version 1.0 if you're ok with it.

If yes, let's create an org. and I'll add you as an maintainer of the npm package.

br