@heruan 2016-06-27
Thank you @matjaz! Can you explain the changes of this PR one by one? I'd like to document it well.