@matjaz 2016-06-27
Hey,

I think this are the main changes:
- extracted `injectFn` for reusage in other decorators
- removed `undefined` comparison (we can check falsy values)
- renamed `InvocationHandlerWrapper` to `InvocationHandler` and removed export in index (not needed)
- in `configure` make a decision if we really need new `InvocationHandler`, otherwise return original one
- in `InvocationHandler` create object using `Object.create`, allowing us to pass properties before constructor is called (removed afterConstructor (not needed))
- unit tests converted to TypeScript
- added autoinject & configure tests