@heruan 2016-06-27
Great! Would you report a simple performance test between the current plugin and your PR to see if using `Object.create` is a plus/plus rather than a plus/minus? In the meantime I'll create the org and prepare the migration!