@matjaz 2016-06-27
Here are some results from simple test


Test case:
```javascript
fdescribe('perf', () => {
    it('test', () => {
        const COUNT = 99999

        @transient('x')
        class App {}
        // comment this out to test non property injected classes
        App.injectProperties = {}

        const container = getContainer()
        console.time('t1')
        for (var i=0; i<COUNT; i++) {
            container.get(App)
        }
        console.timeEnd('t1')
    })
})
```

Results:

Type | without injectProperties (ms) | using injectProperties (ms)
------------ | ------------- | -------
Current master | 91 | 105
PR | 69 | 150

It's expected `Object.create` is slower than `new Fn` since it allows more flexibility.
It's interesting to see mine PR is faster for classes without `injectProperties` (not creating new `InvocationHandler`).
I think props in constructor is a big plus for code maintainability so if performance is really a concern we could add configuration which strategy to use.
What do you think?