@matjaz 2016-06-29
I've added plugin `injectConstructor` config. Also added TypeScript code coverage.