@heruan 2016-07-04
Thank you @matjaz I was oot the last few days, I'll check it out asap and merge.