@heruan 2016-07-07
Merged and moved to a new org, you should have received the invite! I'm already updated the README to point to the NPM installation, please publish version 0.4.0 there and possibly add me as maintainer. Thank you for you help, I look forward to keep this repo updated together and collaborate on other projects too!