@matjaz 2016-07-07
Published 0.4 version and added you as collaborator. 👍